api/requirements.txt
flask
sqlalchemy
flask-sqlalchemy
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
